Bug 1677426 - Part 1: Move MemberInitializers to SharedStencil.h. r=nbp
To avoid including JSScript.h from Stencil.h, moved MemberInitializers out of
JSScript.h

Differential Revision: https://phabricator.services.mozilla.com/D97658
